refactor: 카드 정보 context 관련 코드 파일 분리 및 적용
