+++ src/contexts/cardInfo/CardInfoContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+import { CardInfoContextType } from "./types";
+
+const CardInfoContext = createContext<CardInfoContextType>({
+  cardInfoList: [],
+  dispatch: () => {},
+});
+
+export default CardInfoContext;
