+++ src/contexts/cardInfo/useCardInfo.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import CardInfoContext from "./CardInfoContext";
+
+const useCardInfo = () => {
+  const context = useContext(CardInfoContext);
+
+  return context;
+};
+
+export default useCardInfo;
